pyproject.toml
[project]
name = "Open_Duck_Playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "framesviewer>=1.0.2",
    "jax[cuda12]>=0.5.0",
    "jaxlib>=0.5.0",
    "jaxonnxruntime>=0.3.0",
    "jaxtyping>=0.2.38",
    "matplotlib>=3.10.0",
    "mediapy>=1.2.2",
    "onnxruntime>=1.20.1",
    "playground>=0.0.3",
    "pygame>=2.6.1",
    "tensorflow>=2.18.0",
    "tf2onnx>=1.16.1",
]
[build-system]
requires = ["setuptools"]
[tool.setuptools]
packages = ["playground"]
